refac: put deposit fetching and search filter logic inside a hook
